+++ logs/.24b32864f9f1a8ddd79a0516d1c57bcf29f133a8-audit.json
@@ -9,6 +9,11 @@
             "date": 1769778232647,
             "name": "E:\\Dev-LNT-BE\\namastey-node\\logs\\combined-2026-01-30.log",
             "hash": "ea1c5b6e6863f2b87b0eacfe2a2fe7c13fd3f36e0118b5c61140fd1fe3090385"
+        },
+        {
+            "date": 1769840453595,
+            "name": "E:\\Dev-LNT-BE\\namastey-node\\logs\\combined-2026-01-31.log",
+            "hash": "ae4751c05114445de73a71db9983de280dcc0700da58c3355b8af238253caa80"
         }
     ],
     "hashType": "sha256"
