+++ logs/.411dd6acd3c6a7ceee5cf3152d797f0d370730b0-audit.json
@@ -9,6 +9,11 @@
             "date": 1769778232649,
             "name": "E:\\Dev-LNT-BE\\namastey-node\\logs\\http-2026-01-30.log",
             "hash": "36a20dbf8c9773a77eeb21950dfd67a05b660c8d8692ceb19afe432681b913cd"
+        },
+        {
+            "date": 1769840453597,
+            "name": "E:\\Dev-LNT-BE\\namastey-node\\logs\\http-2026-01-31.log",
+            "hash": "14721623005c4aa8797b7ebc1a134a7f6ad1e2033c787c5a2a590cf174c0f36c"
         }
     ],
     "hashType": "sha256"
