+++ logs/.74a3f62c6d4004180fa7e109eff53d5ec8714d7d-audit.json
@@ -9,6 +9,11 @@
             "date": 1769778232653,
             "name": "E:\\Dev-LNT-BE\\namastey-node\\logs\\database-2026-01-30.log",
             "hash": "8cef3f2aeb634bf027cdaa61f3da041e638c1e4183380f583e35c32a599bda13"
+        },
+        {
+            "date": 1769840453598,
+            "name": "E:\\Dev-LNT-BE\\namastey-node\\logs\\database-2026-01-31.log",
+            "hash": "8bf03cc8f518616820032d1ae4dd6d8c6a4fafbde6dfc6dafa42cb4811e010c7"
         }
     ],
     "hashType": "sha256"
