+++ logs/.ea0a1bd33018fc098e80a0889fc31ca2a34a900a-audit.json
@@ -9,6 +9,11 @@
             "date": 1769778232643,
             "name": "E:\\Dev-LNT-BE\\namastey-node\\logs\\error-2026-01-30.log",
             "hash": "6351a34faa45a70a75afc81f790ed75f24d63aad1ca7a18f994cba2c36f46893"
+        },
+        {
+            "date": 1769840453593,
+            "name": "E:\\Dev-LNT-BE\\namastey-node\\logs\\error-2026-01-31.log",
+            "hash": "bc316fa5e5a6fcc1ef448dfc5f1a95079bc5fe8e0dcc464dde69ace64bbcc3e5"
         }
     ],
     "hashType": "sha256"
